+++ app/controllers/Post.php
@@ -22,19 +22,31 @@ class Post extends Controller
 	
 	public function add()
 	{
+		if ($this->hasPermission()) {
 		$this->post->insert($_POST['title'], $_POST['post_content'], $_SESSION['pseudo']);
 		echo $this->twig->render('home.twig', ['message' => Message::ADDED]);
+		} else {
+			$this->forbidden();
+		}
 	}
 	
 	public function delete()
 	{
+		if ($this->hasPermission()) {
 		$this->post->delete($_GET['id']);
 		echo $this->twig->render('home.twig', ['message' => Message::DELETED_POST]);
+		} else {
+			$this->forbidden();
+		}
 	}
 	
 	public function edit()
 	{
+		if ($this->hasPermission()) {
 		$this->post->edit($_POST['title'], $_POST['message'], $_POST['author'], $_GET['id']);
 		echo $this->twig->render('home.twig', ['message' => Message::EDITED]);
+		} else {
+			$this->forbidden();
+		}
 	}
 }
